bump version to 0.4 and: - add documentation to sentences data structs - change exports of sentences and other types
