+++ Cargo.toml
@@ -14,7 +14,7 @@ license = "MIT OR Apache-2.0"
 name = "nmea"
 readme = "README.md"
 repository = "https://github.com/AeroRust/nmea"
-version = "0.3.1"
+version = "0.4.0"
 
 edition = "2021"
 rust-version = "1.56"
