+++ src/parse.rs
@@ -8,10 +8,25 @@ use nom::{
     IResult,
 };
 
-use crate::{Error, SentenceType};
+use crate::{sentences::*, Error, SentenceType};
 
-pub use crate::sentences::*;
-
+/// The maximum message length parsable by the crate.
+///
+/// From `gpsd`:
+///
+/// > We've had reports that on the Garmin GPS-10 the device sometimes
+/// (1:1000 or so) sends garbage packets that have a valid checksum
+/// but are like 2 successive NMEA packets merged together in one
+/// with some fields lost. Usually these are much longer than the
+/// legal limit for NMEA, so we can cope by just tossing out overlong
+/// packets.  This may be a generic bug of all Garmin chipsets.
+/// NMEA 3.01, Section 5.3 says the max sentence length shall be
+/// 82 chars, including the leading $ and terminating \r\n.
+///
+/// > Some receivers (TN-200, GSW 2.3.2) emit oversized sentences.
+/// The Trimble BX-960 receiver emits a 91-character GGA message.
+/// The current hog champion is the Skytraq S2525F8 which emits
+/// a 100-character PSTI message.
 pub const SENTENCE_MAX_LEN: usize = 102;
 
 /// A known and parsable Nmea sentence type.
@@ -35,7 +50,7 @@ impl<'a> NmeaSentence<'a> {
     }
 }
 
-pub fn checksum<'a, I: Iterator<Item = &'a u8>>(bytes: I) -> u8 {
+pub(crate) fn checksum<'a, I: Iterator<Item = &'a u8>>(bytes: I) -> u8 {
     bytes.fold(0, |c, x| c ^ *x)
 }
 
@@ -51,8 +66,6 @@ fn parse_sentence_type(i: &str) -> IResult<&str, SentenceType> {
     map_res(take(3usize), |sentence_type: &str| {
         SentenceType::try_from(sentence_type).map_err(|_| "Unknown sentence type")
     })(i)
-
-    // .map_err(|_err| Error::Unknown(sentence_type))
 }
 
 fn do_parse_nmea_sentence(i: &str) -> IResult<&str, NmeaSentence> {
@@ -74,22 +87,6 @@ fn do_parse_nmea_sentence(i: &str) -> IResult<&str, NmeaSentence> {
 }
 
 pub fn parse_nmea_sentence(sentence: &str) -> core::result::Result<NmeaSentence, Error<'_>> {
-    // From gpsd:
-    //
-    // We've had reports that on the Garmin GPS-10 the device sometimes
-    // (1:1000 or so) sends garbage packets that have a valid checksum
-    // but are like 2 successive NMEA packets merged together in one
-    // with some fields lost. Usually these are much longer than the
-    // legal limit for NMEA, so we can cope by just tossing out overlong
-    // packets.  This may be a generic bug of all Garmin chipsets.
-    // NMEA 3.01, Section 5.3 says the max sentence length shall be
-    // 82 chars, including the leading $ and terminating \r\n.
-    //
-    // Some receivers (TN-200, GSW 2.3.2) emit oversized sentences.
-    // The Trimble BX-960 receiver emits a 91-character GGA message.
-    // The current hog champion is the Skytraq S2525F8 which emits
-    // a 100-character PSTI message.
-
     if sentence.len() > SENTENCE_MAX_LEN {
         Err(Error::SentenceLength(sentence.len()))
     } else {
@@ -97,6 +94,7 @@ pub fn parse_nmea_sentence(sentence: &str) -> core::result::Result<NmeaSentence,
     }
 }
 
+/// The result of parsing a single NMEA message.
 #[derive(Debug, PartialEq)]
 pub enum ParseResult {
     BOD(BodData),
@@ -110,30 +108,34 @@ pub enum ParseResult {
     RMC(RmcData),
     TXT(TxtData),
     VTG(VtgData),
+    /// A message that is not supported by the crate and cannot be parsed.
     Unsupported(SentenceType),
 }
 
+/// Parse a NMEA 0183 sentence from bytes and extract data from it.
 ///
 /// # Errors
 ///
-/// Apart from errors returned by message parsing itself, it can return additional errors:
-///
-/// - [`Error::Utf8Decoding`] when bytes are not valid UTF-8 string
-/// - [`Error::ASCII`] when bytes are not ASCII only
+/// Apart from errors returned by the message parsing itself, it will return
+/// [`Error::Utf8Decoding`] when the bytes are not a valid UTF-8 string.
 pub fn parse_bytes(sentence_input: &[u8]) -> Result<ParseResult, Error> {
     let string = core::str::from_utf8(sentence_input).map_err(|_err| Error::Utf8Decoding)?;
-
-    if !string.is_ascii() {
-        return Err(Error::ASCII);
-    }
 
     parse_str(string)
 }
 
-/// Parse a NMEA 0183 sentence and extract data from it.
+/// Parse a NMEA 0183 sentence from a string slice and extract data from it.
 ///
 /// Should not contain `\r\n` ending.
+///
+/// # Errors
+///
+/// - [`Error::ASCII`] when string contains non-ASCII characters.
 pub fn parse_str(sentence_input: &str) -> Result<ParseResult, Error> {
+    if !sentence_input.is_ascii() {
+        return Err(Error::ASCII);
+    }
+
     let nmea_sentence = parse_nmea_sentence(sentence_input)?;
     let calculated_checksum = nmea_sentence.calc_checksum();
