+++ src/sentences/mod.rs
@@ -1,4 +1,4 @@
-use nom::{character::complete::anychar, combinator::opt, IResult};
+//! All the supported sentence type data and parsers.
 
 mod bod;
 mod bwc;
@@ -13,164 +13,27 @@ mod txt;
 mod utils;
 mod vtg;
 
-pub use bod::{parse_bod, BodData};
-pub use bwc::{parse_bwc, BwcData};
-pub use gbs::{parse_gbs, GbsData};
-pub use gga::{parse_gga, GgaData};
-pub use gll::{parse_gll, GllData};
-pub use gns::{parse_gns, GnsData, NavigationStatus};
-pub use gsa::{parse_gsa, GsaData};
-pub use gsv::{parse_gsv, GsvData};
-pub use rmc::{parse_rmc, RmcData, RmcStatusOfFix};
-pub use txt::{parse_txt, TxtData};
-pub use vtg::{parse_vtg, VtgData};
+pub(crate) mod faa_mode;
+mod fix_type;
+mod gnss_type;
 
-use crate::FixType;
-
-/// for now let's handle only two GPS and GLONASS
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct FaaModes {
-    sys_state0: FaaMode,
-    sys_state1: Option<FaaMode>,
-}
-
-impl From<FaaModes> for FixType {
-    fn from(modes: FaaModes) -> Self {
-        let fix_type: FixType = modes.sys_state0.into();
-        if fix_type.is_valid() {
-            return fix_type;
-        }
-        if let Some(fix_type2) = modes.sys_state1.map(FixType::from) {
-            if fix_type2.is_valid() {
-                return fix_type2;
-            }
-        }
-        fix_type
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum FaaMode {
-    /// A - Autonomous mode
-    Autonomous,
-    /// C - Quectel Querk, "Caution"
-    Caution,
-    /// D - Differential Mode
-    Differential,
-    /// E - Estimated (dead-reckoning) mode
-    Estimated,
-    /// F - RTK Float mode
-    FloatRtk,
-    /// M - Manual Input Mode
-    Manual,
-    /// N - Data Not Valid
-    DataNotValid,
-    /// P - Precise (4.00 and later)
-    ///
-    /// Sort of DGPS, NMEA 4+
-    Precise,
-    /// R - RTK Integer mode
-    FixedRtk,
-    /// S - Simulated Mode
-    Simulator,
-    /// U - Quectel Querk, "Unsafe"
-    Unsafe,
-}
-
-impl From<FaaMode> for FixType {
-    fn from(mode: FaaMode) -> Self {
-        match mode {
-            FaaMode::Autonomous => FixType::Gps,
-            FaaMode::Caution => FixType::Invalid,
-            FaaMode::Differential => FixType::DGps,
-            FaaMode::Estimated => FixType::Estimated,
-            FaaMode::FloatRtk => FixType::FloatRtk,
-            FaaMode::DataNotValid => FixType::Invalid,
-            FaaMode::Precise => FixType::DGps,
-            FaaMode::FixedRtk => FixType::Rtk,
-            FaaMode::Manual => FixType::Manual,
-            FaaMode::Simulator => FixType::Simulation,
-            FaaMode::Unsafe => FixType::Invalid,
-        }
-    }
-}
-
-pub(crate) fn parse_faa_modes(i: &str) -> IResult<&str, FaaModes> {
-    let (rest, sym) = anychar(i)?;
-
-    let mut ret = FaaModes {
-        sys_state0: parse_faa_mode(sym).ok_or_else(|| nom_parse_failure(i))?,
-        sys_state1: None,
-    };
-
-    let (rest2, sym) = opt(anychar)(rest)?;
-
-    match sym {
-        Some(sym) => {
-            ret.sys_state1 = Some(parse_faa_mode(sym).ok_or_else(|| nom_parse_failure(rest))?);
-        }
-        None => return Ok((rest, ret)),
-    };
-
-    if rest2.is_empty() {
-        Ok((rest2, ret))
-    } else {
-        Err(nom_parse_failure(rest2))
-    }
-}
-
-pub(crate) fn parse_faa_mode(value: char) -> Option<FaaMode> {
-    match value {
-        'A' => Some(FaaMode::Autonomous),
-        'C' => Some(FaaMode::Caution),
-        'D' => Some(FaaMode::Differential),
-        'E' => Some(FaaMode::Estimated),
-        'F' => Some(FaaMode::FloatRtk),
-        'N' => Some(FaaMode::DataNotValid),
-        'P' => Some(FaaMode::Precise),
-        'R' => Some(FaaMode::FixedRtk),
-        'M' => Some(FaaMode::Manual),
-        'S' => Some(FaaMode::Simulator),
-        'U' => Some(FaaMode::Unsafe),
-        _ => None,
-    }
-}
+pub use {
+    bod::{parse_bod, BodData},
+    bwc::{parse_bwc, BwcData},
+    faa_mode::{FaaMode, FaaModes},
+    fix_type::FixType,
+    gbs::{parse_gbs, GbsData},
+    gga::{parse_gga, GgaData},
+    gll::{parse_gll, GllData},
+    gns::{parse_gns, GnsData, NavigationStatus},
+    gnss_type::GnssType,
+    gsa::{parse_gsa, GsaData},
+    gsv::{parse_gsv, GsvData},
+    rmc::{parse_rmc, RmcData, RmcStatusOfFix},
+    txt::{parse_txt, TxtData},
+    vtg::{parse_vtg, VtgData},
+};
 
 pub(crate) fn nom_parse_failure(inp: &str) -> nom::Err<nom::error::Error<&str>> {
     nom::Err::Failure(nom::error::Error::new(inp, nom::error::ErrorKind::Fail))
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_parse_faa_modes() {
-        assert_eq!(
-            nom::Err::Error(nom::error::Error::new("", nom::error::ErrorKind::Eof)),
-            parse_faa_modes("").unwrap_err(),
-            "Should return a Digit error on empty string"
-        );
-        assert_eq!(
-            (
-                "",
-                FaaModes {
-                    sys_state0: FaaMode::Autonomous,
-                    sys_state1: None,
-                }
-            ),
-            parse_faa_modes("A").unwrap()
-        );
-
-        assert_eq!(
-            (
-                "",
-                FaaModes {
-                    sys_state0: FaaMode::DataNotValid,
-                    sys_state1: Some(FaaMode::Autonomous),
-                }
-            ),
-            parse_faa_modes("NA").unwrap()
-        );
-    }
 }
